requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
docker==6.1.3
geographiclib==2.0
geopy==2.4.0
idna==3.4
packaging==23.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.7
websocket-client==1.6.4